Add save/loading csv/json persistence. Other changes and improvements.

+++ Data/Abstract/IProgress.cs
@@ -4,6 +4,6 @@ namespace LastIRead {
     public interface IProgress {
         DateTime Date { get; }
 
-        int Value { get; }
+        double Value { get; }
     }
 }
